Added Meeting Notes from 9/28/2018
During this meeting, we narrowed down our requirements. We were able to start communicating new ideas based on these requirements and develop new questions for Bassam. Additionally, we started thinking about what forms of technology we will use. Once we narrow down on the technology we want to use, we can start modeling an interface to prototype our idea.